requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
httpx==0.25.2
beautifulsoup4==4.12.2
newspaper3k==0.2.8
python-multipart==0.0.6
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
email-validator==2.1.0
jinja2==3.1.2
aiofiles==23.2.1
python-dateutil==2.8.2
pyyaml==6.0.1
markdownify==0.11.6